requirements.txt
firecrawl-py>=1.0.0
crawl4ai>=0.4.0
openai>=1.0.0
anthropic>=0.40.0
pydantic>=2.0.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
markdownify>=0.13.0
trafilatura>=1.8.0
openpyxl>=3.1.0
pandas>=2.0.0
